Remove avx512_BW_VL (#460)
* remove bw_vl arch (not needed, avx512 is enough)

* remove xenial from merge queue flow

* fix fit memory test (#461)

* fix fit memory test

* remove xenial checks

* remove irrelevant comment

* free index in seralization test

* install gcc10 bionic

* install newer vg ubuntu18

* exit valgrind directory
