+++ .install/ubuntu_18.04.sh
@@ -9,6 +9,15 @@ $MODE apt-get dist-upgrade -yqq
 $MODE apt install -yqq software-properties-common
 $MODE add-apt-repository ppa:ubuntu-toolchain-r/test -y
 $MODE apt update
-$MODE apt-get install -yqq git wget make valgrind gcc-9 g++-9
+$MODE apt-get install -yqq git wget make gcc-9 g++-9 libc6-dbg
 $MODE update-alternatives --install /usr/bin/gcc gcc /usr/bin/gcc-9 60 --slave /usr/bin/g++ g++ /usr/bin/g++-9
+
+wget https://sourceware.org/pub/valgrind/valgrind-3.18.0.tar.bz2
+tar -xjf valgrind-3.18.0.tar.bz2
+cd valgrind-3.18.0
+./configure
+make
+$MODE make install
+cd ..
+
 source install_cmake.sh $MODE
