+++ src/VecSim/spaces/CMakeLists.txt
@@ -19,7 +19,6 @@ if(CMAKE_HOST_SYSTEM_PROCESSOR MATCHES "(x86_64)|(AMD64|amd64)|(^i.86$)")
 	# from gcc14+ -mavx512bw is implicitly enabled when -mavx512vbmi2 is requested
 	CHECK_CXX_COMPILER_FLAG(-mavx512bw CXX_AVX512BW)
 	CHECK_CXX_COMPILER_FLAG(-mavx512vbmi2 CXX_AVX512VBMI2)
-	CHECK_CXX_COMPILER_FLAG(-mavx512vl CXX_AVX512VL)
 	CHECK_CXX_COMPILER_FLAG(-mavx512f CXX_AVX512F)
 	CHECK_CXX_COMPILER_FLAG(-mavx2 CXX_AVX2)
 	CHECK_CXX_COMPILER_FLAG(-mavx CXX_AVX)
@@ -37,17 +36,9 @@ if(CMAKE_HOST_SYSTEM_PROCESSOR MATCHES "(x86_64)|(AMD64|amd64)|(^i.86$)")
 		add_compile_definitions(OPT_AVX512_BW_VBMI2)
 	endif()
 
-	if(CXX_AVX512BW AND CXX_AVX512VL)
-		message("Building with AVX512BW and AVX512VL")
-		set_source_files_properties(functions/AVX512BW_VL.cpp PROPERTIES
-				COMPILE_FLAGS "-mavx512f -mf16c -mavx512bw -mavx512vl")
-		list(APPEND OPTIMIZATIONS functions/AVX512BW_VL.cpp)
-		add_compile_definitions(OPT_AVX512_BW_VL)
-	endif()
-
 	if(CXX_AVX512F)
 		message("Building with AVX512")
-		set_source_files_properties(functions/AVX512.cpp PROPERTIES COMPILE_FLAGS "-mavx512f -mf16c")
+		set_source_files_properties(functions/AVX512.cpp PROPERTIES COMPILE_FLAGS "-mavx512f")
 		list(APPEND OPTIMIZATIONS functions/AVX512.cpp)
 		add_compile_definitions(OPT_AVX512F)
 	endif()
