+++ src/VecSim/spaces/functions/AVX512.cpp
@@ -12,6 +12,9 @@
 #include "VecSim/spaces/IP/IP_AVX512_FP32.h"
 #include "VecSim/spaces/IP/IP_AVX512_FP64.h"
 
+#include "VecSim/spaces/IP/IP_AVX512_FP16.h"
+#include "VecSim/spaces/L2/L2_AVX512_FP16.h"
+
 namespace spaces {
 
 #include "implementation_chooser.h"
@@ -37,6 +40,18 @@ dist_func_t<float> Choose_FP32_L2_implementation_AVX512(size_t dim) {
 dist_func_t<double> Choose_FP64_L2_implementation_AVX512(size_t dim) {
     dist_func_t<double> ret_dist_func;
     CHOOSE_IMPLEMENTATION(ret_dist_func, dim, 8, FP64_L2SqrSIMD8_AVX512);
+    return ret_dist_func;
+}
+
+dist_func_t<float> Choose_FP16_IP_implementation_AVX512(size_t dim) {
+    dist_func_t<float> ret_dist_func;
+    CHOOSE_IMPLEMENTATION(ret_dist_func, dim, 32, FP16_InnerProductSIMD32_AVX512);
+    return ret_dist_func;
+}
+
+dist_func_t<float> Choose_FP16_L2_implementation_AVX512(size_t dim) {
+    dist_func_t<float> ret_dist_func;
+    CHOOSE_IMPLEMENTATION(ret_dist_func, dim, 32, FP16_L2SqrSIMD32_AVX512);
     return ret_dist_func;
 }
