+++ src/VecSim/spaces/functions/AVX512.h
@@ -16,4 +16,7 @@ dist_func_t<double> Choose_FP64_IP_implementation_AVX512(size_t dim);
 dist_func_t<float> Choose_FP32_L2_implementation_AVX512(size_t dim);
 dist_func_t<double> Choose_FP64_L2_implementation_AVX512(size_t dim);
 
+dist_func_t<float> Choose_FP16_IP_implementation_AVX512(size_t dim);
+dist_func_t<float> Choose_FP16_L2_implementation_AVX512(size_t dim);
+
 } // namespace spaces
