+++ tests/benchmark/CMakeLists.txt
@@ -36,7 +36,6 @@ if(CMAKE_HOST_SYSTEM_PROCESSOR MATCHES "(x86_64)|(AMD64|amd64)|(^i.86$)")
 	# Check that the compiler supports instructions flag.
 	# This will add the relevant flag both the the space selector and the optimization.
 	CHECK_CXX_COMPILER_FLAG(-mavx512bw CXX_AVX512BW)
-	CHECK_CXX_COMPILER_FLAG(-mavx512vl CXX_AVX512VL)
 	CHECK_CXX_COMPILER_FLAG(-mavx512vbmi2 CXX_AVX512VBMI2)
 	CHECK_CXX_COMPILER_FLAG(-mavx512f CXX_AVX512F)
 	CHECK_CXX_COMPILER_FLAG(-mavx2 CXX_AVX2)
@@ -52,10 +51,6 @@ if(CMAKE_HOST_SYSTEM_PROCESSOR MATCHES "(x86_64)|(AMD64|amd64)|(^i.86$)")
 
 	if(CXX_AVX512BW AND CXX_AVX512VBMI2)
 		add_compile_definitions(OPT_AVX512_BW_VBMI2)
-	endif()
-
-	if(CXX_AVX512BW AND CXX_AVX512VL)
-		add_compile_definitions(OPT_AVX512_BW_VL)
 	endif()
 
 	if(CXX_F16C AND CXX_FMA AND CXX_AVX)
