+++ tests/benchmark/spaces_benchmarks/bm_spaces_fp16.cpp
@@ -12,10 +12,10 @@
 cpu_features::X86Features opt = cpu_features::GetX86Info().features;
 
 // AVX512_BW_VL functions
-#ifdef OPT_AVX512_BW_VL
-bool avx512_bw_vl_supported = opt.avx512bw && opt.avx512vl;
-INITIALIZE_BENCHMARKS_SET(FP16, AVX512BW_VL, 32, avx512_bw_vl_supported);
-#endif // OPT_AVX512_BW_VL
+#ifdef OPT_AVX512F
+bool avx512_supported = opt.avx512f;
+INITIALIZE_BENCHMARKS_SET(FP16, AVX512, 32, avx512_supported);
+#endif // OPT_AVX512F
 
 // AVX functions
 #ifdef OPT_F16C
